+++ wazo_confd/plugins/extension/resource.py
@@ -1,9 +1,7 @@
-# Copyright 2016-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from flask import request, url_for
-
-from xivo_dao.alchemy.extension import Extension
 
 from wazo_confd.auth import required_acl
 from wazo_confd.helpers.restful import ListResource, ItemResource
@@ -13,11 +11,14 @@ from .schema import ExtensionSchema
 
 class ExtensionList(ListResource):
 
-    model = Extension
     schema = ExtensionSchema
 
+    def __init__(self, service, middleware):
+        super().__init__(service)
+        self._middleware = middleware
+
     def build_headers(self, extension):
-        return {'Location': url_for('extensions', id=extension.id, _external=True)}
+        return {'Location': url_for('extensions', id=extension['id'], _external=True)}
 
     @required_acl('confd.extensions.read')
     def get(self):
@@ -25,11 +26,12 @@ class ExtensionList(ListResource):
 
     @required_acl('confd.extensions.create')
     def post(self):
-        form = self.schema().load(request.get_json())
-        model = self.model(**form)
         tenant_uuids = self._build_tenant_list({'recurse': True})
-        model = self.service.create(model, tenant_uuids)
-        return self.schema().dump(model), 201, self.build_headers(model)
+        resource = self._middleware.create(
+            request.get_json(),
+            tenant_uuids,
+        )
+        return resource, 201, self.build_headers(resource)
 
     def _has_a_tenant_uuid(self):
         # The base function does not work because the tenant_uuid is not part